requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==4.3.1
pycryptodome==3.20.0
boto3
gunicorn
django-storages==1.14.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
PyJWT==2.8.0
emoji-country-flag==1.3.2
frozenlist==1.4.1
geoip2==4.8.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
maxminddb==2.5.1
multidict==6.0.4
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.2
coverage==7.4.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3
psycopg2
python-decouple==3.8
wheel==0.41.2
yarl==1.9.4
cryptography==42.0.2
